backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["senanradjabov <iamtheemperor06@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
pydantic-settings = "^2.6.1"
pyjwt = "^2.10.1"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
bcrypt = "4.0.1"
cryptography = "^44.0.0"
python-multipart = "^0.0.19"
python-slugify = "^8.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
